requirements.txt
absl-py==1.4.0
aiohttp==3.8.0
aiosignal==1.3.1
albumentations==0.4.3
altair==4.2.2
antlr4-python3-runtime==4.9.3
anyio==3.6.2
argon2-cffi==21.3.0
argon2-cffi-bindings==21.2.0
arrow==1.2.3
astor==0.8.1
asttokens==2.2.1
async-timeout==4.0.2
attrs==22.2.0
azure-core==1.26.3
backcall==0.2.0
backports.zoneinfo==0.2.1
base58==2.1.1
beautifulsoup4==4.12.0
black==23.3.0
bleach==6.0.0
blinker==1.5
blis==0.7.9
brotlipy==0.7.0
cachetools==5.3.0
cairocffi==1.5.0
CairoSVG==2.7.0
catalogue==2.0.8
certifi==2022.12.7
cffi==1.15.1
chardet==5.1.0
charset-normalizer==2.1.1
click==8.1.3
cloudpickle==2.2.1
coloredlogs==15.0.1
comm==0.1.3
confection==0.0.4
contourpy==1.0.7
cssselect2==0.7.0
cycler==0.11.0
cymem==2.0.7
debugpy==1.6.6
decorator==5.1.1
defusedxml==0.7.1
diffusers==0.12.1
dpcpp-cpp-rt==2023.0.0
einops==0.3.0
emoji==2.2.0
entrypoints==0.4
executing==1.2.0
fastjsonschema==2.16.3
filelock==3.9.0
flatbuffers==23.3.3
fonttools==4.39.0
fqdn==1.5.1
frozenlist==1.3.3
fsspec==2023.3.0
ftfy==6.1.1
future==0.18.3
fvcore==0.1.5.post20221221
gitdb==4.0.10
GitPython==3.1.31
google-auth==2.16.2
google-auth-oauthlib==0.4.6
grpcio==1.51.3
huggingface-hub==0.12.1
humanfriendly==10.0
hydra-core==1.3.2
idna==3.4
imageio==2.9.0
imageio-ffmpeg==0.4.2
imgaug==0.2.6
importlib-metadata==6.0.0
importlib-resources==5.12.0
install==1.3.5
intel-cmplr-lib-rt==2023.0.0
intel-cmplr-lic-rt==2023.0.0
intel-opencl-rt==2023.0.0
intel-openmp==2023.0.0
invisible-watermark==0.1.5
iopath==0.1.9
ipykernel==6.22.0
ipython==8.12.0
ipython-genutils==0.2.0
ipywidgets==8.0.6
isoduration==20.11.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.2.0
jsonpointer==2.3
jsonschema==4.17.3
jupyter==1.0.0
jupyter-console==6.6.3
jupyter-events==0.6.3
jupyter_client==8.1.0
jupyter_core==5.3.0
jupyter_server==2.5.0
jupyter_server_terminals==0.4.4
jupyterlab-pygments==0.2.2
jupyterlab-widgets==3.0.7
kiwisolver==1.4.4
kornia==0.6.0
langcodes==3.3.0
lazy_loader==0.1
lxml==4.9.2
Markdown==3.4.1
MarkupSafe==2.1.2
matplotlib==3.7.1
matplotlib-inline==0.1.6
mistune==2.0.5
mkl==2023.1.0
mkl-fft==1.3.1
mkl-service==2.4.0
mpmath==1.2.1
multidict==6.0.4
murmurhash==1.0.9
mypy-extensions==1.0.0
nbclassic==0.5.4
nbclient==0.7.3
nbconvert==7.3.0
nbformat==5.8.0
nest-asyncio==1.5.6
networkx==3.0
nltk==3.8.1
notebook==6.5.3
notebook_shim==0.2.2
numpy==1.22.4
nvidia-cublas-cu11==11.10.3.66
nvidia-cuda-nvrtc-cu11==11.7.99
nvidia-cuda-runtime-cu11==11.7.99
nvidia-cudnn-cu11==8.5.0.96
oauthlib==3.2.2
omegaconf==2.3.0
onnx==1.13.1
onnxruntime==1.14.1
openai==0.27.0
opencv-python==4.1.2.30
opencv-python-headless==4.7.0.72
packaging==23.0
pandas==1.5.3
pandocfilters==1.5.0
parso==0.8.3
pathspec==0.11.1
pathy==0.10.1
pexpect==4.8.0
pickleshare==0.7.5
Pillow==9.4.0
pkgutil_resolve_name==1.3.10
platformdirs==3.2.0
portalocker==2.7.0
preshed==3.0.8
prometheus-client==0.16.0
prompt-toolkit==3.0.38
protobuf==3.20.3
psutil==5.9.4
ptyprocess==0.7.0
pudb==2019.2
pure-eval==0.2.2
pyarrow==11.0.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycocotools==2.0.6
pycparser==2.21
pydantic==1.10.7
pydeck==0.8.0
pyDeprecate==0.3.1
Pygments==2.14.0
pyparsing==3.0.9
pyrallis==0.3.1
pyrsistent==0.19.3
python-dateutil==2.8.2
python-json-logger==2.0.7
pytorch-lightning==1.4.2
pytz==2022.7.1
pytz-deprecation-shim==0.1.0.post0
PyWavelets==1.4.1
PyYAML==6.0
pyzmq==25.0.2
qtconsole==5.4.2
QtPy==2.3.1
regex==2022.10.31
reportlab==3.6.12
requests==2.28.2
requests-oauthlib==1.3.1
rfc3339-validator==0.1.4
rfc3986-validator==0.1.1
rsa==4.9
SceneGraphParser==0.1.0
scikit-image==0.20.0
scipy==1.9.1
seaborn==0.12.2
Send2Trash==1.8.0
six==1.16.0
smart-open==6.3.0
smmap==5.0.0
sniffio==1.3.0
soupsieve==2.4
spacy==3.5.1
spacy-legacy==3.0.12
spacy-loggers==1.0.4
srsly==2.4.6
stack-data==0.6.2
stanza==1.5.0
streamlit==0.73.1
svglib==1.5.1
sympy==1.11.1
tabulate==0.9.0
tbb==2021.9.0
tensorboard==2.12.0
tensorboard-data-server==0.7.0
tensorboard-plugin-wit==1.8.1
termcolor==2.2.0
terminado==0.17.1
test-tube==0.7.5
thinc==8.1.9
tifffile==2023.2.28
timm==0.6.12
tinycss2==1.2.1
tokenizers==0.12.1
toml==0.10.2
tomli==2.0.1
toolz==0.12.0
torch-fidelity==0.3.0
torchaudio==0.13.1
torchmetrics==0.6.0
tornado==6.2
tqdm==4.65.0
traitlets==5.9.0
transformers==4.22.1
typer==0.7.0
typing-inspect==0.8.0
typing_extensions==4.5.0
tzdata==2022.7
tzlocal==4.2
uri-template==1.2.0
urllib3==1.26.15
urwid==2.1.2
validators==0.20.0
Wand==0.6.11
wasabi==1.1.1
watchdog==2.3.1
wcwidth==0.2.6
webcolors==1.13
webencodings==0.5.1
websocket-client==1.5.1
Werkzeug==2.2.3
widgetsnbextension==4.0.7
yacs==0.1.8
yarl==1.8.2
zipp==3.15.0
